Move AllocatorInstance to its own header
